requirements.txt
# File: requirements.txt
watchdog==2.1.9
gitpython==3.1.30
python-dotenv>=1.0.0,<2.0.0
chainlit==1.0.0
pygments